框架运行起来 #1

+++ src/main/java/com/reaier/engking/domain/SourceWord.java
@@ -28,7 +28,7 @@ public class SourceWord extends Auditable<Integer> implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonProperty("id")
     @Column(name = "id",                        columnDefinition = "INT UNSIGNED")
-    Long id;
+    Integer id;
 
     @ApiModelProperty(notes = "源主键")
     @JsonProperty("source_id")
@@ -38,5 +38,5 @@ public class SourceWord extends Auditable<Integer> implements Serializable {
     @ApiModelProperty(notes = "单词主键")
     @JsonProperty("word_id")
     @Column(name = "word_id",                   columnDefinition = "INT UNSIGNED NOT NULL COMMENT '单词主键'")
-    Long wordId;
+    Integer wordId;
 }
